chore(clippy): 清理 workspace clippy warnings 至 0
- cargo clippy --fix 自动修复 58 个文件的机械问题（冗余闭包/多余转换/无用生命周期等）

- 手动修复：needless_range_loop / explicit_counter_loop 改迭代器、manual_clamp、ptr_arg、field_reassign_with_default、文档注释、删除死代码 make_theme、mod tests 移到文件末尾

- synth.rs 的 bind group 循环保留索引形式并加 allow+注释：必须固定迭代 MAX_CHUNKS 次，sc 不足时其余 slot 用 dummy buffer 占位

- 修复 view_math.rs 恒真断言（clippy deny 级 absurd_extreme_comparisons）

- 全量测试 402 个通过

+++ crates/yinhe-egui/src/right_panel/channels_panel.rs
@@ -88,14 +88,14 @@ pub fn show(ui: &mut egui::Ui, doc: Option<&mut Document>, settings: &AudioSetti
         .body(|body| {
             body.rows(18.0, compacted_channels as usize, |mut row| {
                 let d = row.index();
-                let sources = &reverse[d as usize];
+                let sources = &reverse[d];
                 let source_label = if sources.is_empty() {
                     "—".to_string()
                 } else {
                     sources
                         .iter()
                         .map(|&src| {
-                            let port = (src >> 4) as u8;
+                            let port = src >> 4 ;
                             let ch = (src & 0x0F) + 1;
                             format!(
                                 "{}{:02}",
